Update to Rails 7 with jsbundling-rails
The project was old enough that many of the dependencies just didn't
build any more. Updating to the modern versions of things meant that
they work since they're still supported.

This also removes a few unused dependencies, like Bourbon and
CoffeeScript. It also updates some deprecated uses of the Rails version
of `Enumerable#sum`.

Because of amcharts3, we can't use importmap-rails. It's not properly
packaged by jspm.io and is missing the dataloader module.